made page back in to server and made client component
